+++ src/FolkerKinzel.Strings/md/ReleaseNotes.md
@@ -42,4 +42,11 @@ string TrimEnd(this string, string?);
 string TrimStart(this string, string?);
 ```
 &nbsp;
+- New Polyfills for the StringBuilder class (.NET Standard 2.0, .NET Framework 4.6.1):
+```csharp
+StringBuilder Trim(this StringBuilder, string?);
+StringBuilder TrimEnd(this StringBuilder, string?);
+StringBuilder TrimStart(this StringBuilder, string?);
+```
+&nbsp;
 > **Project reference:** On some systems, the content of the CHM file in the Assets is blocked. Before opening the file right click on the file icon, select Properties, and **check the "Allow" checkbox** - if it is present - in the lower right corner of the General tab in the Properties dialog.